+++ lib/maze/hooks/appium_hooks.rb
@@ -6,6 +6,7 @@ module Maze
       @client
 
       def before_all
+        Maze::Plugins::DatadogMetricsPlugin.send_increment('appium.test_started')
         @client = Maze::Client::Appium.start
       end
 
@@ -30,6 +31,11 @@ module Maze
 
       def after_all
         @client&.log_run_outro
+        if $success
+          Maze::Plugins::DatadogMetricsPlugin.send_increment('appium.test_succeeded')
+        else
+          Maze::Plugins::DatadogMetricsPlugin.send_increment('appium.test_failed')
+        end
       end
 
       def at_exit